Pull functions out into separate file
