+++ src/js/components/editListing.mjs
@@ -1,6 +1,12 @@
 import { API_BASE_URL } from "../api/baseUrl.js";
 import { getSpecificListing } from "../pages/specificListing.js";
 import { clearMessages, displayMessage } from "../utils/displayMessage.mjs";
+import {
+  createAndInsertImageField,
+  isValidURL,
+  canInsertNewImage,
+  maxNumberOfImages,
+} from "../utils/imageInputs.js";
 import { token } from "../utils/storage.mjs";
 
 const queryString = window.location.search;
@@ -10,9 +16,6 @@ const editFormContainer = document.querySelector("#edit-listing-form");
 const saveButton = document.querySelector("#save-listing-button");
 const imageInputContainer = document.querySelector("#editImageInputsContainer");
 const addImageInputButton = document.querySelector("#addImageInput");
-
-// Image Input Limit
-const maxNumberOfImages = 5;
 
 // Get values
 async function getValues() {
@@ -41,9 +44,9 @@ async function getValues() {
     if (json.media.length === 0) {
       createAndInsertImageField("", imageInputContainer);
     } else {
-      json.media.forEach((image) =>
-        createAndInsertImageField(image, imageInputContainer),
-      );
+      json.media.forEach((image) => {
+        createAndInsertImageField(image, imageInputContainer);
+      });
     }
   } catch (error) {
     console.log(error);
@@ -86,57 +89,6 @@ addImageInputButton.addEventListener("click", () => {
   createAndInsertImageField("", imageInputContainer);
 });
 
-// Function to check if a string is a valid URL
-function isValidURL(url) {
-  try {
-    new URL(url);
-    return true;
-  } catch (error) {
-    return false;
-  }
-}
-
-function canInsertNewImage() {
-  // Check if number of images is within range.
-  const imageInputs = document.querySelectorAll(".image-input.form-control");
-  return imageInputs.length < maxNumberOfImages;
-}
-
-function createAndInsertImageField(image, container) {
-  // Create a new row for image input
-  const newRow = document.createElement("div");
-  newRow.className = "image-input-row";
-
-  const flexImgContainer = document.createElement("div");
-  flexImgContainer.className = "d-flex";
-
-  // Create input for image URL
-  const newImageInput = document.createElement("input");
-  newImageInput.type = "url";
-  newImageInput.className = "image-input form-control";
-  newImageInput.value = image;
-
-  // Create delete button for the new image input
-  const deleteImageUrlButton = document.createElement("i");
-  deleteImageUrlButton.className = "fa-regular fa-trash-can";
-  deleteImageUrlButton.id = "delete-image-input";
-  deleteImageUrlButton.style.color = "#272c35";
-  deleteImageUrlButton.addEventListener("click", () => {
-    newRow.remove();
-    if (canInsertNewImage()) {
-      clearMessages(".empty-image-message");
-    }
-  });
-
-  // Append the elements to the new row
-  newRow.appendChild(flexImgContainer);
-  flexImgContainer.appendChild(newImageInput);
-  flexImgContainer.appendChild(deleteImageUrlButton);
-
-  // Append the new row to the container
-  container.appendChild(newRow);
-}
-
 // Edit values
 editFormContainer.addEventListener("submit", (e) => {
   e.preventDefault();
@@ -173,8 +125,6 @@ async function editListing(newTitle, newDescription, newMedia, newTags) {
 
   try {
     const response = await fetch(putUrl, putOptions);
-    const json = await response.json();
-    console.log(json);
 
     if (response.ok) {
       displayMessage(
